requirements.txt
amqp==5.2.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
billiard==4.2.2
black==24.10.0
blinker==1.9.0
brotli==1.2.0
celery==5.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
ConfigArgParse==1.7.1
coverage==7.11.3
cron_descriptor==2.0.6
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.0.6
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-graphql-jwt==0.4.0
django-redis==5.4.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-nested-routers==0.93.5
drf-postman==0.1.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
drf-yasg==1.21.8
executing==2.2.1
flake8==7.1.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
flower==2.0.1
gevent==25.9.1
geventhttpclient==2.3.5
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
humanize==4.14.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
ipython==8.28.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.4.2
locust==2.31.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
model-bakery==1.20.5
msgpack==1.1.2
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
promise==2.3
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.23
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==5.0.1
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.28.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
whitenoise==6.6.0
zope.event==6.1
zope.interface==8.1
